Disable seccomp
It does not work on armhf Ubuntu 16.04.  I suspect the kernel was not
built with support, but I'm not sure.  But this brings up a bigger
issue: I would need to verify every architecture before turning this on.
And in any event, it can't be enabled on kFreeBSD (or Hurd, but ntpd
does not work on Hurd anyway).