Update the refclock list
This commit disables the oncore and gpsd refclocks.  The oncore was
deprecated upstream for being broken.

Additional refclocks were deprecated upstream for only returning two
digit years.  That seems like a tractable problem, given the build
epoch.  Since I use the spectracom driver, I'm going to look into that.

gpsd is a tough call.  It's not deprecated upstream, and it still can
work.  However, upstream recommends the use of SHM, as does every
example HOWTO I found with a quick search.  Furthermore, it is buggy:
  https://lists.ntpsec.org/pipermail/devel/2016-October/002392.html
Also, it "makes all sorts of non-standard assumptions about how gpsd is
configured.  A configuration that is contorted and non-obvious":
  https://lists.ntpsec.org/pipermail/devel/2016-October/002392.html
Most importantly, it is an attractive nuisance.  If someone is told that
the best practice is to use "gpsd", they might logically think they
should use the gpsd driver.