Make ntpwait less strict
We do not need full synchronization before time-sync.target.  The goal
is to ensure that the time is close, but most importantly, not going to
move backwards.  (This is all that is provided by systemd-timesyncd, for
example.)  The initial stepping of the clock is sufficient.  This change
can mean the difference of many minutes in delay.