Fix ntploggps systemd depdendencies w/o gpsd
Instead of having timers.target pull in the timers, we have the relevant
daemon server (ntpd.service or gpsd.service) pull in the timers.  This
way, if the daemon is not running, the timers don't run either.  This is
especially important for gpsd.service.  If one doesn't have gpsd
installed, ntploggps.timer / ntploggps.service would fail.