Improve debian/copyright
On 02/18/2018 08:55 AM, Chris Lamb <lamby@debian.org> wrote:
> I just ACCEPTed ntpsec from NEW but noticed it was missing
> attribution in debian/copyright for at least Motorolo,
> autorevision.sh, Chris Johns, etc.

Motorola is not missing. The "COPYRIGHT 1991-1994 MOTOROLA INC." and
similar in ntpd/refclock_oncore.c are part of a copy of text from a
label on hardware, to illustrate the models of hardware the driver was
tested with. I have now added this as a comment in debian/copyright.

I have specifically addressed autorevision.sh and Chris Johns.

I also made another pass over everything, and found a few more missing.

Closes: 890758
Reported-by: Chris Lamb <lamby@debian.org>