Create /var/lib/ntp in the postinst
This way, the ntp user and group are already created, and we avoid the
need for a chown.  The chown had a -R, which Lintian now frowns upon for
security reasons.  The -R was probably unnecessary anyway, but if
needed, we can always bring back a chown for ntp.drift.