Fix installing ntpsec-ntpviz without ntpsec
When installing ntpsec-ntpviz without ntpsec, the installation would
fail because the ntp user and group did not exist.  The simple fix is to
copy the adduser call from ntpsec into ntpsec-ntpviz.

The cleanup is much more complicated.  Both packages will remove the ntp
user and group when *purged*, but only if the other package is also
gone.  The defintion of "gone" here varies between ntpsec and
ntpsec-ntpviz.  For ntpsec-ntpviz, "gone" means it has been or is being
removed.  For ntpsec, "gone" means it has been or is being purged.  The
difference is because the ntpsec package does not recursively delete
/var/log/ntpstats until it is purged.  Those files would be owned by the
ntp user and group.

Also, since ntpsec-ntpviz does not require ntpsec, it can be installed
alongside the ntp package, so it also requires ntp to be "gone".

Closes: 891278
Reported-by: Andreas Beckmann <anbe@debian.org>