Stop deleting /var/lib/ntpdate/
The ntpdate package used to use /var/lib/ntpdate and deletes it in the
postrm.  The ntpsec-ntpdate package inherited that in the fork.  ntpdate
has since stopped using it entirely (using /run/ntpdate.dhcp instead of
/var/lib/ntpdate/default.dhcp).  That change was merged into the
ntpsec-ntpdate package before 1.0.0-1, so there has never been a
released version of ntpsec-ntpdate using /var/lib/ntpdate.  Accordingly,
the easiest solution for us is to simply stop removing it.

Closes: 892966
Reported-by: Julian Gilbey <jdg@debian.org>