Use ntp.conf.dhcp if it exists
The old test was "if it is newer than /etc/ntp.conf".  It was suggested,
in Debian bug #898402, that I make this change, as the admin might
change /etc/ntp.conf, restart, have it work, and then be surprised when
the next DHCP renewal overrides their settings again.

This change brings ntpsec in line with ntp.  It also brings the ntpsec
sysvinit script in line with the ntpsec systemd wrapper.

I looked through the git history, and I can't find a good reason for why
this "-nt" test existed. It's unclear if that came from an older version
of ntp or was something I cooked up.