debian/copyright: Merge in full CC0 text
I already had the full CC-BY-4.0 text in debian/copyright.  I had the
full CC0 text in debian/COPYING-CC0.txt, but it probably belongs in
debian/copyright.  The machine-readable specification says, "this field
should either include the full text of the license(s) or include a
pointer to the license file under /usr/share/common-licenses".  It is
not available in /usr/share/common-licenses.