ntpdate-debian: Do nothing if called from ntp's ntpdate hook
If ntp's ntpdate package has been removed but not purged, it will leave
an /etc/network/if-up.d/ntpdate hook which calls ntpdate-debian.  We do
not need nor want to sync in that case, as that might lead to syncing
twice instead of once or once instead of none, depending on whether the
user has ntpsec-ntpdate installed.

Gbp-Dch: Ignore