Add net_admin to the apparmor profile
Neither I nor upstream is completely sure of why this is necessary.  Our
hunch is that it's related to the interface routing code.  I am only
able to reproduce the denial message once per boot, but it's unclear if
that is related to the logging or if the denial really only occurs once.