Drop dependency on python3-all-dev
The ntpsec package is only building for the default Python version, not
all Python versions.  At this point, the only applications using the
"ntp" module are the client utilities in this package, so there is very
little point in trying to build for jython or ironpython.  Also, if
upstream ends up moving to Go, the "ntp" python module will disappear.

Closes: 909731