ntpdate.8: Some minor corrections to the manual
Remove space characters at the end of lines.

Use the correct macro for the font change of a single argument.

19:.IR server

Change - to \- if it shall be printed as a minus sign.

69:call, even if the measured offset is greater than +-128 ms. The
71:is greater than +-128 ms. Note that, if the offset is much
72:greater than +-128 ms in this case, that it can take a long

Remove a repeated word

! 119 --> is

Split lines longer than 80 characters into two or more lines.
Appropriate break points are the end of a sentence and a subordinate
clause; after punctuation marks

ntpdate.8: line 27	length 106
the correct time. It must be run as root on the local host (unless the
option \fB\-q\fR is used). A number

Use \(en for a dash (en-dash) between space characters, not a minus
(\-) or a hyphen (-), except in the NAME section.

110:Query only - don't set the clock.

Write "man page" or "manual page", see man-pages(7), not "manpage".

Closes: 915937
Signed-off-by: Richard Laager <rlaager@wiktel.com>