Merge in ntpdate.8 changes from ntp
This is the "Update ntpdate.8 from ntpdate.html" commit:
https://salsa.debian.org/pkg-ntp-team/ntp/commit/9fed56816201051a806203f129302d904029079a

I preserved the change to -B (because this ntpdate, at least, always
slews by default).  I kept the fix to the duplicated "is" in -t.  I also
kept the "1 second" (not 2 seconds, as the ntp commit changed it to), as
that is the value used by this ntpdate.

Thanks: Bernhard Schmidt <berni@debian.org>