Set maxclock in the default config
The default maxclock setting is 10.  The pool entries count (which I
don't like, but that's been the behavior forever), so this is really 6
server entries.  It's better that the number of servers be an odd
number, as 2n+1 servers protect against n falsetickers.  Alternatively,
we could go down one to 5 servers by setting maxclock to 9.