Commit changelog
I'm committing the changelog so that I can see if this makes a
difference for the Salsa pipeline.  It looks like the piuparts job fails
if the changelog has not been updated since the last upload, as it sees
a version conflict between the package in the archive and the built
package.

Gbp-Dch: Ignore